:green_heart: Add needed variable exports
Needed because of how OPPO's fingerprint drivers work. They expect the kernel to always be built as a part of AOSP, where TARGET_PRODUCT (and more) would be defined. The makefile uses these to define variables inside the driver source to remove sections of code on some devices. Because we're not building as part of a ROM, we need to export these manually within the script.